+++ src/main/java/com/efs/sdk/organizationmanager/core/auth/OrgaOwnerController.java
@@ -63,20 +63,39 @@ public class OrgaOwnerController {
         return ResponseEntity.ok(ownerService.listOwners(authHelper.getAuthenticationModel(token), orgaId));
     }
 
-    @Operation(summary = "Sets owners in Organization by userIds", description = """
-            Sets all owners in `Organization` by userIds.
-                        
-            **CAUTION** overwrites the list of owners - so the list has to be complete""")
+    @Operation(summary = "Assigns Ownership to a Specific Organization", description = """
+            This endpoint assigns a new set of owners to the specified `Organization`. Owners can be set using either their user IDs or email addresses, based on the specified `type` parameter.
+                       
+            Supported `type` values:
+             - `email`: The list of owners contains email addresses.
+             - `userId`: The list of owners contains user IDs.
+                       
+            *Note*: The operation will not proceed if any provided user IDs or email addresses are invalid or do not exist.
+            
+            **CAUTION** overwrites the list of owners - the provided list should be comprehensive and complete.
+            """)
     @PutMapping(produces = MediaType.APPLICATION_JSON_VALUE)
-    @ApiResponse(responseCode = "200", description = "Successfully set owners by user-ids.")
-    @ApiResponse(responseCode = "403", description = "User does not have permission to set owners", content = @Content(schema = @Schema(hidden = true)))
-    @ApiResponse(responseCode = "404", description = "Organization or User was not found.", content = @Content(schema = @Schema(hidden = true)))
-    public ResponseEntity<OrganizationReadDTO> setOwners(@Parameter(hidden = true) JwtAuthenticationToken token, @PathVariable @Parameter(description = "The "
-            + "id of the `Organization`.") long orgaId, @RequestBody List<String> owners) throws OrganizationmanagerException {
-        Organization organization = ownerService.setOwners(authHelper.getAuthenticationModel(token), orgaId, owners);
+    @ApiResponse(responseCode = "200", description = "Owners successfully set")
+    @ApiResponse(responseCode = "403", description = "Insufficient permission to set owners.", content = @Content(schema = @Schema(hidden = true)))
+    @ApiResponse(responseCode = "404", description = "Either `Organization` or `User` not found.", content = @Content(schema = @Schema(hidden = true)))
+    public ResponseEntity<OrganizationReadDTO> setOwners(
+            @Parameter(hidden = true) JwtAuthenticationToken token,
+            @PathVariable @Parameter(description = "The id of the `Organization`.") long orgaId,
+            @RequestParam(required = false, defaultValue = "userId") String type,
+            @RequestBody List<String> owners) throws OrganizationmanagerException {
+        Organization organization;
+        if (type.equalsIgnoreCase("userId")) {
+            organization = ownerService.setOwners(authHelper.getAuthenticationModel(token), orgaId, owners);
+        } else if (type.equalsIgnoreCase("email")) {
+            organization = ownerService.setOwnersByEmail(authHelper.getAuthenticationModel(token), orgaId, owners);
+        } else {
+            throw new OrganizationmanagerException(OrganizationmanagerException.ORGANIZATIONMANAGER_ERROR.BAD_REQUEST_PARAM_VALUE, String.format("-- type " +
+                    "'%s' does not exist", type));
+        }
         OrganizationReadDTO orgaDTO = converter.convertToDTO(organization, OrganizationReadDTO.class);
         return ResponseEntity.ok(orgaDTO);
     }
+
 
     @Operation(summary = "Adds owner in Organization by username", description = "Adds owner in `Organization` by username")
     @PutMapping(path = "name/{name}", produces = MediaType.APPLICATION_JSON_VALUE)